Comment update method

+++ Finance.API/Controllers/CommentsController.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using Finance.API.Dtos.Comment;
+using Finance.API.Dtos.Stock;
 using Finance.API.Interfaces;
 using Finance.API.Models;
 using Microsoft.AspNetCore.Http;
@@ -25,24 +26,30 @@ namespace Finance.API.Controllers
 		[HttpGet]
 		public async Task<IActionResult> GetAllAsync()
 		{
+			if (!ModelState.IsValid) return BadRequest(ModelState);
+			
 			var comments =  await _commentRepository.GetAllAsync();
 			var map = _mapper.Map<List<CommentDto>>(comments);
 			return Ok(map);	
 		}
 
 
-		[HttpGet("{id}")]
+		[HttpGet("{id:int}")]
 		public async Task<IActionResult> GetByIdAsync(int id)
 		{
+			if (!ModelState.IsValid) return BadRequest(ModelState);
+
 			var comment = await _commentRepository.GetByIdAsync(id);
 			var map = _mapper.Map<CommentDto>(comment);
 			if (map == null) return NotFound();
 			return Ok(map);
 		}
 
-		[HttpPost("{stockId}")]
+		[HttpPost("{stockId:int}")]
 		public async Task<IActionResult> CreateComment([FromRoute] int stockId, CreateCommentRequestDto createCommentRequestDto)
 		{
+			if (!ModelState.IsValid) return BadRequest(ModelState);
+
 			var stock = await _stockRepository.StockExist(stockId);
 			if (stockId != createCommentRequestDto.StockId)
 			{
@@ -59,12 +66,23 @@ namespace Finance.API.Controllers
 
 			return Ok(commentDto); // Bu satırı düzeltin
 		}
+		[HttpPut]
+		[Route("{id:int}")]
+		public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
+		{
+			if (!ModelState.IsValid)
+				return BadRequest(ModelState);
 
+			var commentModel = await _commentRepository.UpdateAsync(id, updateDto);
+			if (commentModel == null) return NotFound();
+			return Ok(updateDto);
+		}
 
 		[HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
 		public async Task<IActionResult> Delete([FromRoute] int id)
 		{
+			if (!ModelState.IsValid) return BadRequest(ModelState);
 			await _commentRepository.DeleteAsync(id);
 			return NoContent();
 		}
